requirements.txt
packaging
six
ydb
ydb[yc]
pyTelegramBotAPI
python-dotenv